# Conflicts:
#	src/server/game/Handlers/PetBattleHandler.cpp

+++ src/server/game/Guilds/Guild.h
@@ -473,11 +473,11 @@ class Guild
                     return IsMoneyEvent(m_eventType);
                 }
 
-                BankEventLogEntry(uint32 guildId, uint32 guid, GuildBankEventLogTypes eventType, uint8 tabId, uint32 playerGuid, uint32 itemOrMoney, uint16 itemStackCount, uint8 destTabId) :
+                BankEventLogEntry(uint32 guildId, uint32 guid, GuildBankEventLogTypes eventType, uint8 tabId, uint32 playerGuid, uint64 itemOrMoney, uint16 itemStackCount, uint8 destTabId) :
                     LogEntry(guildId, guid), m_eventType(eventType), m_bankTabId(tabId), m_playerGuid(playerGuid),
                     m_itemOrMoney(itemOrMoney), m_itemStackCount(itemStackCount), m_destTabId(destTabId) { }
 
-                BankEventLogEntry(uint32 guildId, uint32 guid, time_t timestamp, uint8 tabId, GuildBankEventLogTypes eventType, uint32 playerGuid, uint32 itemOrMoney, uint16 itemStackCount, uint8 destTabId) :
+                BankEventLogEntry(uint32 guildId, uint32 guid, time_t timestamp, uint8 tabId, GuildBankEventLogTypes eventType, uint32 playerGuid, uint64 itemOrMoney, uint16 itemStackCount, uint8 destTabId) :
                     LogEntry(guildId, guid, timestamp), m_eventType(eventType), m_bankTabId(tabId), m_playerGuid(playerGuid),
                     m_itemOrMoney(itemOrMoney), m_itemStackCount(itemStackCount), m_destTabId(destTabId) { }
 
@@ -490,7 +490,7 @@ class Guild
                 GuildBankEventLogTypes m_eventType;
                 uint8  m_bankTabId;
                 uint32 m_playerGuid;
-                uint32 m_itemOrMoney;
+                uint64 m_itemOrMoney;
                 uint16 m_itemStackCount;
                 uint8  m_destTabId;
         };
@@ -908,7 +908,7 @@ class Guild
         bool _MemberHasTabRights(uint64 guid, uint8 tabId, uint32 rights) const;
 
         void _LogEvent(GuildEventLogTypes eventType, uint32 playerGuid1, uint32 playerGuid2 = 0, uint8 newRank = 0);
-        void _LogBankEvent(SQLTransaction& trans, GuildBankEventLogTypes eventType, uint8 tabId, uint32 playerGuid, uint32 itemOrMoney, uint16 itemStackCount = 0, uint8 destTabId = 0);
+        void _LogBankEvent(SQLTransaction& trans, GuildBankEventLogTypes eventType, uint8 tabId, uint32 playerGuid, uint64 itemOrMoney, uint16 itemStackCount = 0, uint8 destTabId = 0);
 
         Item* _GetItem(uint8 tabId, uint8 slotId) const;
         void _RemoveItem(SQLTransaction& trans, uint8 tabId, uint8 slotId);
