+++ src/server/game/Scripting/ScriptMgr.h
@@ -348,7 +348,7 @@ class ScriptMgr
         /// @p_ItemOrMoney    : Item entry or gold amount
         /// @p_ItemStackCount : Item stack count
         /// @p_DestTabID      : Destination tab ID
-        void OnGuildBankEvent(Guild* p_Guild, uint8 p_EventType, uint8 p_TabID, uint32 p_PlayerGUID, uint32 p_ItemOrMoney, uint16 p_ItemStackCount, uint8 p_DestTabID);
+        void OnGuildBankEvent(Guild* p_Guild, uint8 p_EventType, uint8 p_TabID, uint32 p_PlayerGUID, uint64 p_ItemOrMoney, uint16 p_ItemStackCount, uint8 p_DestTabID);
 
     /// ItemScript
     public:
